+++ example-nextjs/helpers/useDrag.ts
@@ -3,17 +3,16 @@ import React from 'react';
 export default function useDrag() {
   const [clicked, setClicked] = React.useState(false);
   const [dragging, setDragging] = React.useState(false);
-  const [position, setPosition] = React.useState(0);
-  const [diff, setDiff] = React.useState(0);
+  const position = React.useRef(0);
 
   const dragStart = React.useCallback((ev: React.MouseEvent) => {
-    setPosition(ev.clientX);
-    setDiff(0);
+    position.current = ev.clientX;
     setClicked(true);
   }, []);
 
   const dragStop = React.useCallback(
     () =>
+      // NOTE: need some delay so item under cursor won't be clicked
       window.requestAnimationFrame(() => {
         setDragging(false);
         setClicked(false);
@@ -21,34 +20,27 @@ export default function useDrag() {
     []
   );
 
-  const dragMove = React.useCallback(
-    (ev: React.MouseEvent, cb: (newPos: number) => void) => {
-      const newDiff = position - ev.clientX;
+  const dragMove = (ev: React.MouseEvent, cb: (posDiff: number) => void) => {
+    const newDiff = position.current - ev.clientX;
 
-      const movedEnough = Math.abs(newDiff) > 5;
+    const movedEnough = Math.abs(newDiff) > 5;
 
-      if (clicked && movedEnough) {
-        setDragging(true);
-      }
+    if (clicked && movedEnough) {
+      setDragging(true);
+    }
 
-      if (dragging && movedEnough) {
-        setPosition(ev.clientX);
-        setDiff(newDiff);
-        cb(newDiff);
-      }
-    },
-    [clicked, dragging, position]
-  );
+    if (dragging && movedEnough) {
+      position.current = ev.clientX;
+      cb(newDiff);
+    }
+  };
 
   return {
     dragStart,
     dragStop,
     dragMove,
-    diff,
     dragging,
     position,
     setDragging,
-    setDiff,
-    setPosition,
   };
 }