+++ src/ItemsMap.ts
@@ -1,12 +1,22 @@
 import type { IOItem, Item, visibleItems as itemsArr } from './types';
+import { filterSeparators } from './helpers';
 
 class ItemsMap extends Map<string, IOItem> {
   public toArr(): Item[] {
     return [...this];
   }
 
-  public toItemsKeys(): itemsArr {
+  public toItems(): itemsArr {
     return this.toArr().map((el) => el[0]);
+  }
+
+  public toItemsWithoutSeparators(): itemsArr {
+    return filterSeparators(this.toItems());
+  }
+
+  // NOTE: for backward compatibility, to remove
+  public toItemsKeys(): itemsArr {
+    return this.toItems();
   }
 
   onlyDigits(value: string | number): string {
@@ -29,6 +39,7 @@ class ItemsMap extends Map<string, IOItem> {
     }
     return this;
   }
+
   public first(): IOItem | undefined {
     return this.toArr()[0]?.[1];
   }