+++ src/index.tsx
@@ -12,6 +12,9 @@ import { observerOptions as defaultObserverOptions } from './settings';
 
 import * as constants from './constants';
 
+import useOnInitCb from './hooks/useOnInitCb';
+import useOnUpdate from './hooks/useOnUpdate';
+
 import { VisibilityContext } from './context';
 
 import type { ItemType, Refs } from './types';
@@ -21,12 +24,12 @@ import slidingWindow from './slidingWindow';
 import getItemsPos from './getItemsPos';
 
 type ArrowType = React.FC | React.ReactNode;
-
-interface Props {
+export interface Props {
   LeftArrow?: ArrowType;
   RightArrow?: ArrowType;
   children: ItemType | ItemType[];
   onInit?: (api: publicApiType) => void;
+  onUpdate?: (api: publicApiType) => void;
   onScroll?: (api: publicApiType, ev: React.UIEvent) => void;
   onWheel?: (api: publicApiType, ev: React.WheelEvent) => void;
   options?: Partial<typeof defaultObserverOptions>;
@@ -37,6 +40,7 @@ interface Props {
   separatorClassName?: string;
   scrollContainerClassName?: string;
   wrapperClassName?: string;
+  apiRef?: React.MutableRefObject<publicApiType>;
 }
 
 function ScrollMenu({
@@ -44,6 +48,7 @@ function ScrollMenu({
   RightArrow: _RightArrow,
   children,
   onInit = (): void => void 0,
+  onUpdate = (): void => void 0,
   onMouseDown,
   onMouseUp,
   onMouseMove,
@@ -54,6 +59,7 @@ function ScrollMenu({
   itemClassName = '',
   separatorClassName = '',
   wrapperClassName = '',
+  apiRef = { current: {} as publicApiType },
 }: Props): JSX.Element {
   const LeftArrow = getElementOrConstructor(_LeftArrow);
   const RightArrow = getElementOrConstructor(_RightArrow);
@@ -74,7 +80,9 @@ function ScrollMenu({
   // NOTE: hack for detect when items added/removed dynamicaly
   const itemsChanged = useItemsChanged(children, menuItemsRefs);
 
-  const items = React.useRef(new ItemsMap()).current;
+  const _items = React.useRef(new ItemsMap());
+  const items = _items.current;
+
   const { visibleItems } = useIntersectionObserver({
     items,
     itemsChanged,
@@ -84,31 +92,34 @@ function ScrollMenu({
 
   const publicApi = React.useRef<publicApiType>({} as publicApiType);
 
-  const initComplete = !!visibleItems.length;
-  const [onInitFired, setOnInitFired] = React.useState(false);
-  React.useEffect(() => {
-    if (initComplete && !onInitFired) {
-      setOnInitFired(true);
-      onInit(publicApi.current);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [initComplete, onInitFired]);
+  const mounted = !!visibleItems.length;
+
+  const onInitCbFired = useOnInitCb({
+    cb: () => onInit(publicApi.current),
+    condition: mounted,
+  });
+
+  useOnUpdate({
+    cb: () => onUpdate(publicApi.current),
+    condition: onInitCbFired,
+    visibleItems,
+  });
 
   const api = React.useMemo(
     () => createApi(items, visibleItems),
     [items, visibleItems]
   );
-
   publicApi.current = React.useMemo(
     () => ({
       ...api,
-      initComplete,
+      initComplete: mounted,
       items,
       scrollContainer: scrollContainerRef,
-      visibleItems,
     }),
-    [api, initComplete, items, visibleItems]
+    [api, mounted, items]
   );
+
+  apiRef.current = publicApi.current;
 
   const scrollHandler = React.useCallback(
     (event: React.UIEvent) => onScroll(publicApi.current, event),